% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/banco.sim.3PL.R
\docType{data}
\name{banco.sim.3PL}
\alias{banco.sim.3PL}
\title{Banco simulado simples}
\format{
An object of class \code{list} of length 2.
}
\usage{
banco.sim.3PL
}
\description{
Banco simulado simples
}
\details{
Lista com dois elementos:
\itemize{
\item \code{respostas} banco com 3.000 respostas e 45 itens de múltipla escolha. As respostas
possíveis são "A", "B", "C", "D", "E", ".", "*"
\item \code{gabarito} gabarito dos 45 itens
}
}
\keyword{datasets}
