% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/banco.sim.BIB.aberto.R
\docType{data}
\name{banco.sim.BIB.aberto}
\alias{banco.sim.BIB.aberto}
\title{Banco com BIB aberto simulado}
\format{
An object of class \code{list} of length 4.
}
\usage{
banco.sim.BIB.aberto
}
\description{
Banco com BIB aberto simulado
}
\details{
Lista com dois elementos:
\itemize{
\item \code{respostas} banco com 10.000 respostas e 77 itens de múltipla escolha. As respostas
possíveis são "A", "B", "C", "D", "E", ".", "*"
\item \code{gabarito} gabarito dos 77 itens
\item \code{BIB} desenho do BIB
\item \code{itens} dicionário dos itens
}
}
\keyword{datasets}
