% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brincar2.R
\name{brincar2}
\alias{brincar2}
\title{Brincando com a TRI}
\usage{
brincar2()
}
\description{
Gera um aplicativo shiny com o objetivo de conhecer melhor a TRI
}
\examples{
brincar2()
}
\author{
Alexandre Jaloto
}
