% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcular_bisserial.R
\name{calcular_bisserial}
\alias{calcular_bisserial}
\title{Calcular correlação bisserial alterada}
\usage{
calcular_bisserial(resp.item, escore, gabarito, pop)
}
\arguments{
\item{resp.item}{objeto com as respostas ao item}

\item{escore}{escore dos sujeitos}

\item{gabarito}{gabarito do item}

\item{pop}{TRUE se for população, FALSE (padrão) se for uma amostra. Essa escolha interfere nas contas que envolvem o cálculo da
variância ou do desvio padrão.}
}
\value{
Valor da correlação bisserial alterada
}
\description{
Calcula correlação bisserial quando a variável dicotômica
é tratada como contínua com uma distribuição normal subjacente
}
\author{
Alexandre Jaloto
}
\keyword{internal}
