% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcular_normit.R
\name{calcular_normit}
\alias{calcular_normit}
\title{Calcular normit}
\usage{
calcular_normit(banco.aberto, tot.cad, escore)
}
\arguments{
\item{banco.aberto}{banco aberto obtido em \code{abre.banco}}

\item{tot.cad}{máximo de cadernos}

\item{escore}{escore dos sujeitos}
}
\value{
Vetor numérico com os valores de normit dos sujeitos
}
\description{
Calcula o valor de normit para os sujeitos quando eles
são submetidos a formas diferentes de um teste
}
\author{
Alexandre Jaloto
}
\keyword{internal}
