% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcular_proporcao.R
\name{calcular_proporcao}
\alias{calcular_proporcao}
\title{Calcular proporção de acerto}
\usage{
calcular_proporcao(resp.item, gabarito)
}
\arguments{
\item{resp.item}{objeto com as respostas ao item}

\item{gabarito}{gabarito do item}
}
\value{
Valor da proporção de acerto ao item
}
\description{
Calcula a proporção de acerto a um item
}
\author{
Alexandre Jaloto
}
\keyword{internal}
