% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compara.sim.tct.R
\name{compara.sim.tct}
\alias{compara.sim.tct}
\title{Comparar análise pela TCT de bancos simulados}
\usage{
compara.sim.tct(banco, resultado)
}
\arguments{
\item{banco}{lista com o banco simulado (respostas e gabarito)}

\item{resultado}{objeto com resultado da TCT gerado pela função \code{tct}}
}
\value{
A função mostra a comparação na tela.
}
\description{
Compara resultados da análise pela TCT de bancos simulados.
}
\examples{
compara.sim.tct (banco = banco.sim.3PL, resultado = tct.3PL)

}
\author{
Alexandre Jaloto
}
