% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compara.sim.tri.v1.R
\name{compara.sim.tri.v1}
\alias{compara.sim.tri.v1}
\title{Comparar rodada V1 da análise pela TRI de bancos simulados}
\usage{
compara.sim.tri.v1(banco, tab.pars, objeto.mirt)
}
\arguments{
\item{banco}{lista com o banco simulado (respostas e gabarito)}

\item{tab.pars}{tabela de parâmetros gerado pela função \code{mirt} ou \code{multipleGroup} com argumento \code{pars = 'values'}}

\item{objeto.mirt}{objeto mirt da calibração}
}
\value{
A função mostra a comparação na tela.
}
\description{
Compara resultados da V1 da análise pela TRI de bancos simulados.
}
\examples{
compara.sim.tri.v1 (banco = banco.sim.3PL, tab.pars = values, objeto.mirt = fit1)

}
\author{
Alexandre Jaloto
}
