% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compara.tabelas.R
\name{compara.tabelas}
\alias{compara.tabelas}
\title{Comparar a tabela gerada com a simulada}
\usage{
compara.tabelas(objeto.insumo, objeto.sim)
}
\arguments{
\item{objeto.insumo}{tabela gerada pelo usuário}

\item{objeto.sim}{tabela oficial gerada com dados simulados}
}
\value{
Vetor lógico que indica existência de diferença (\code{TRUE}) ou ausência (\code{FALSE}).
}
\description{
Compara uma tabela gerada pelo usuário com a tabela oficial gerada com dados simulados
}
\author{
Alexandre Jaloto
}
