% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dif.mirt.R
\name{dif.mirt}
\alias{dif.mirt}
\title{Análise de DIF para o mirt}
\usage{
dif.mirt(
  fit.antigo,
  fit.atual,
  comuns.antigo = NULL,
  comuns.atual = NULL,
  int.teta = c(-6, 6),
  n.qdpt = 61
)
}
\arguments{
\item{fit.antigo}{objeto mirt do tipo \code{SingleGroupClass} referente ao grupo em que os itens foram calibrados (grupo de referência).}

\item{fit.atual}{objeto mirt do tipo \code{SingleGroupClass} referente ao grupo da aplicação em voga (grupo focal).}

\item{comuns.antigo}{vetor com nomes dos itens comuns no banco antigo. Os nomes dos itens podem ser obtidos com \code{mirt::extract.mirt(fit.antigo, 'itemnames')}. Só é necessário se os nomes dos itens forem diferentes nos dois bancos.}

\item{comuns.atual}{vetor com nomes dos itens comuns no banco atual Os nomes dos itens podem ser obtidos com \code{mirt::extract.mirt(fit.atual, 'itemnames')}. Só é necessário se os nomes dos itens forem diferentes nos dois bancos.}

\item{n.qdpt}{quantidade de pontos de quadratura.}

\item{int.qdpt}{intervalo dos pontos de quadratura. Esse intervalo será usado para a análise de MaxADif. O padrão é \code{c(-6, 6)}.}
}
\value{
A função retorna uma lista com quatro elementos
\itemize{
\item \code{itens} vetor com os nomes dos itens que apresentaram DIF. Se os nomes nos bancos forem diferentes, os nomes do banco atual são apresentados
\item \code{maxadif} \code{data.frame} com os itens que paresentaram DIF pelo método MaxADIF, com o nome do item, o ponto de quadratura em que a proporção superou \code{0.15} e as diferenças em cada categoria de resposta
\item \code{regressao} lista com os itens que apresentaram DIF uniforme, não uniforme e misto com o método de regressão logística
\item \code{rmsd} \code{data.frame} com os itens que paresentaram DIF pelo método RMSD, com o nome do item e os valores de RMSD para cada categoria de resposta
}
}
\description{
Verifica existência de DIF e a qualidade do ajuste de um item calibrado no mirt
}
\examples{

set.seed(1234)

a <- rlnorm(60)
d <- rnorm(60)
data.atual <- data.frame(simdata(a, d, 1000, '2PL'))
data.antigo <- data.frame(simdata(a, d, 1000, '2PL'))

names(data.antigo) <- c(paste0('IME_', 1:50), paste0('IRC_', 51:60))
names(data.atual) <- c(paste0('IME_', 61:85), paste0('IME_', 1:25), paste0('IRC_', 51:60))

for(i in 51:60)
{
 data.antigo[,i] <- sample(3, 1000, TRUE)
 data.atual[,i] <- sample(3, 1000, TRUE)
}
fit.antigo <- mirt(data.antigo, 1, TOL = .01)
fit.atual <- mirt(data.atual, 1, TOL = .01)
dif <- dif.mirt(fit.antigo, fit.atual)

# para itens com nomes diferentes
names(data.antigo) <- paste0('I', 1:ncol(data.antigo))

comuns.antigo <- c(paste0('I', 1:25), paste0('I', 51:60))
comuns.atual <- c(paste0('IME_', 1:25), paste0('IRC_', 51:60))

fit.antigo <- mirt(data.antigo, 1, TOL = .01)
fit.atual <- mirt(data.atual, 1, TOL = .01)

dif <- dif.mirt(fit.antigo, fit.atual, comuns.antigo = comuns.antigo, comuns.atual = comuns.atual)

}
\author{
Alexandre Jaloto
}
