% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtro.par.R
\name{filtro.par}
\alias{filtro.par}
\title{Filtrar itens pelos parâmetros}
\usage{
filtro.par(pars, a = c(0.5, 4), b = c(-2.5, 2.5), c = 0.45)
}
\arguments{
\item{pars}{\code{data.frame} com os parâmetros dos itens}

\item{a}{vetor com valores mínimo e máximo do intervalo aceitável para o parâmetro \code{a}}

\item{b}{vetor com valores mínimo e máximo do intervalo aceitável para o parâmetro \code{b}}

\item{c}{valor máximo aceitável para o parâmetro \code{c}}
}
\value{
A função mostra a comparação na tela.
}
\description{
Filtra os itens de acordo com os intervalos estabelecidos para os parâmetros da TRI
}
\details{
Para a análise, são excluídos automaticamente os itens que apresentem pelo menos um dos seguintes critérios:
\itemize{
\item a < .5 ou a > 4
\item nenhum parâmetro b entre -2.5 e 2.5
\item c > .45
\item Problema de ajuste
\item Existência de DIF
}
}
\examples{
compara.sim.tri.v1 (banco = banco.sim.3PL, tab.pars = values, objeto.mirt = fit1)

}
\author{
Alexandre Jaloto
}
