% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gera.banco.compara.R
\name{gera.banco.compara}
\alias{gera.banco.compara}
\title{Gerar banco para comparação}
\usage{
gera.banco.compara(banco)
}
\arguments{
\item{banco}{lista com o banco simulado (respostas e gabarito)}
}
\value{
Banco aberto
}
\description{
Gera o banco aberto para as análises de TCT e TRI dos bancos simulados
}
\examples{
gera.banco.compara (banco = banco.sim.3PL)


}
\author{
Alexandre Jaloto
}
