% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agi.R
\name{grafico.agi}
\alias{grafico.agi}
\title{Grar gráficos AGI}
\usage{
grafico.agi(banco, gabarito, escore, cortes)
}
\arguments{
\item{banco}{\code{data.frame} cuja primeira variável é o \code{ID} dos sujeitos
e as demais são as respostas aos itens.}

\item{gabarito}{\code{data.frame} cuja primeira variável é o nome dos
itens e a segunda, o gabarito do item}

\item{escore}{vetor com os escores dos sujeitos}

\item{cortes}{pontos de corte para classificar os sujeitos de acordo
com o escore}
}
\value{
Retorna uma lista nomeada de objetos da classe \code{ggplot2}.
Cada elemento da lista é um gráfico que representa a Análise Gráfica
do Item (AGI) para um item específico. Os nomes dos elementos da
lista correspondem aos números dos itens (ex: \code{lista_graficos[[1]]}
contém o gráfico do item 1).
}
\description{
Gerar gráficos para Análise Gráfica do Item (AGI)
}
\examples{

data(banco.sim.3PL)
banco <- banco.sim.3PL$respostas
gabarito <- banco.sim.3PL$gabarito
correcao <- mirt::key2binary(banco[,-1], gabarito$Gabarito)
escore <- rowSums(correcao, na.rm = TRUE)
grafico <- grafico.agi(banco = banco, gabarito = gabarito, escore = escore, cortes = seq(0, 45, 9))
plot(grafico[[1]])

}
\author{
Alexandre Jaloto
}
