% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pars.priori.R
\name{pars.priori}
\alias{pars.priori}
\title{Altera priori dos parâmetros}
\usage{
pars.priori(values)
}
\arguments{
\item{values}{tabela de parâmetros gerado pela função \code{mirt} ou \code{multipleGroup} com argumento \code{pars = 'values'}}
}
\value{
Tabela de parâmetros com valores alterados
}
\description{
Altera valores iniciais e distribuição prévia dos parâmetros dos itens na tabela de parâmetros
}
\details{
Distribuição do \code{a} é \code{lognormal} com média \code{0.53062825106217}, que
corresponde a \code{log(1.7)}, e desvio \code{0.5},
com valor inicial \code{1.7}. Distribuição do \code{c} é \code{beta} com alpha \code{5} e beta \code{17},
com valor inicial \code{0.2}.
}
\examples{
pars.priori (parametros)

}
\author{
Alexandre Jaloto
}
