% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/passo_e.R
\name{passo_e}
\alias{passo_e}
\title{Probabilidade esperada}
\usage{
passo_e(a, b, c, nodes = seq(-4, 4, length.out = 40), resp, mu = 0, sigma = 1)
}
\arguments{
\item{a}{valor do parâmetro a}

\item{b}{valor do parâmetro b}

\item{c}{valor do parâmetro c}

\item{nodes}{pontos de quadratura}

\item{resp}{padrão de resposta dos sujeitos}

\item{mu}{média do grupo}

\item{sigma}{desvio padrão do grupo}
}
\value{
Lista com quantidade esperada de sujeitos em cada
ponto de quadratura e proporção esperada de acertos por item em
cada ponto
}
\description{
Gera as probabilidades esperadas (etapa E do algoritmo EM)
}
\author{
Alexandre Jaloto
}
