% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/banco.sim.BIB.fechado.R
\docType{data}
\name{banco.sim.BIB.fechado}
\alias{banco.sim.BIB.fechado}
\title{Banco com BIB fechado simulado}
\format{
An object of class \code{list} of length 4.
}
\usage{
banco.sim.BIB.fechado
}
\description{
Banco com BIB fechado simulado
}
\details{
Lista com quatro elementos:
\itemize{
\item \code{respostas} \code{data.frame} banco de respostas com três variáveis:
\itemize{
\item \code{ID} identificação dos 10.000 sujeitos
\item \code{CADERNO} caderno que o sujeito respondeu
\item \code{TX_RESPOSTA} vetor de resposta aos 77 itens. As respostas
possíveis são "A", "B", "C", "D", "E", ".", "*"
}
\item \code{gabarito} gabarito dos 77 itens
\item \code{BIB} desenho do BIB
\item \code{itens} dicionário dos itens
}
}
\keyword{datasets}
