% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dif.mirt.R
\name{dif.mirt}
\alias{dif.mirt}
\title{Análise de ajuste e DIF em um objeto mirt}
\usage{
dif.mirt(
  fit.atual,
  fit.antigo = NULL,
  comuns.atual = NULL,
  comuns.antigo = NULL,
  int.teta = c(-6, 6),
  n.qdpt = 61,
  int.cent = TRUE,
  limite.max.dif = 0.15,
  limite.rmsd = 0.1,
  limite.rmsd.pisa = 0.1,
  p.ajuste = c(0.05, 0.95),
  p.dif = c(0.05, 0.95),
  limite.pseudoR2.lordif = 0.035
)
}
\arguments{
\item{fit.atual}{objeto mirt do tipo \code{SingleGroupClass} referente ao grupo da aplicação em voga (grupo focal).}

\item{fit.antigo}{objeto mirt do tipo \code{SingleGroupClass} referente ao grupo em que os itens foram calibrados (grupo de referência).}

\item{comuns.atual}{vetor com nomes dos itens comuns no banco atual. Os nomes dos itens podem ser obtidos com \code{mirt::extract.mirt(fit.atual, 'itemnames')}. Só é necessário se os nomes dos itens forem diferentes nos dois bancos.}

\item{comuns.antigo}{vetor com nomes dos itens comuns no banco antigo. Os nomes dos itens podem ser obtidos com \code{mirt::extract.mirt(fit.antigo, 'itemnames')}. Só é necessário se os nomes dos itens forem diferentes nos dois bancos.}

\item{int.teta}{intervalo dos pontos de quadratura. Esse intervalo será usado para a análise de MaxADif. O padrão é \code{c(-6, 6)}.}

\item{n.qdpt}{quantidade de pontos de quadratura.}

\item{int.cent}{lógico. Se TRUE, o intervalo dos pontos de quadratura é centrado. Se FALSE, o intervalo começa no ponto de quadratura. O padrão é TRUE.}

\item{limite.max.dif}{limite superior para as diferenças absolutas em MaxADIF. Itens com diferenças maiores que este valor são sinalizados. O padrão é \code{0.15}.}

\item{limite.rmsd}{limite superior para o RMSD (ajuste e DIF). Itens com RMSD maior que este valor são sinalizados. O padrão é \code{0.10}.}

\item{limite.rmsd.pisa}{limite superior para o RMSD PISA. Itens com RMSD maior que este valor são sinalizados. O padrão é \code{0.10}.}

\item{limite.pseudoR2.lordif}{limite superior para o pseudo R2 de Nagelkerke na análise de regressão logística (lordif). O padrão é \code{0.035}.}
}
\value{
A função retorna uma lista com quatro elementos
\itemize{
\item \code{info} lista com informação usada nos cálculos
\itemize{
\item \code{p.antigo.dif} percentis do grupo antigo para análise de DIF
\item \code{p.atual.dif} percentis do grupo atual para análise de DIF
\item \code{p.ajuste} percentis do grupo atual para análise de ajuste
\item \code{qdpt.analise.dif} pontos de quadratura utilizados na análise de DIF
}
\item \code{itens} vetor com os nomes dos itens que apresentaram DIF. Se os nomes nos bancos forem diferentes, os nomes do banco atual são apresentados
\item \code{rmsd.ajuste} \code{data.frame} com os itens que apresentaram desajuste pelo método RMSD, com o nome do item e os valores de RMSD para cada categoria de resposta
\item \verb{rmsd.pisa``data.frame} com os itens que apresentaram desajuste pelo método RMSD do PISA, com o nome do item e os valores de RMSD para cada categoria de resposta
\item \code{maxadif} \code{data.frame} com os itens que apresentaram DIF pelo método MaxADIF, com o nome do item, o ponto de quadratura em que a proporção superou \code{limite.max.dif} e as diferenças em cada categoria de resposta
\item \code{rmsd.dif} \code{data.frame} com os itens que apresentaram DIF pelo método RMSD, com o nome do item e os valores de RMSD para cada categoria de resposta
\item \code{regressao} lista com os itens que apresentaram DIF uniforme, não uniforme e misto com o método de regressão logística
}
}
\description{
Verifica existência de DIF e a qualidade do ajuste de um item calibrado no mirt
}
\examples{

set.seed(1234)

a <- rlnorm(60)
d <- rnorm(60)
data.atual <- data.frame(mirt::simdata(a, d, 1000, '2PL'))
data.antigo <- data.frame(mirt::simdata(a, d, 1000, '2PL'))

names(data.antigo) <- c(paste0('IME_', 1:50), paste0('IRC_', 51:60))
names(data.atual) <- c(paste0('IME_', 61:85), paste0('IME_', 1:25), paste0('IRC_', 51:60))

for(i in 51:60)
{data.antigo[,i] <- sample(3, 1000, TRUE)
 data.atual[,i] <- sample(3, 1000, TRUE)}
fit.antigo <- mirt::mirt(data.antigo, 1, TOL = .01)
fit.atual <- mirt::mirt(data.atual, 1, TOL = .01)
dif <- dif.mirt(fit.antigo = fit.antigo, fit.atual = fit.atual)

# para itens com nomes diferentes
names(data.antigo) <- paste0('I', 1:ncol(data.antigo))

comuns.antigo <- c(paste0('I', 1:25), paste0('I', 51:60))
comuns.atual <- c(paste0('IME_', 1:25), paste0('IRC_', 51:60))

fit.antigo <- mirt::mirt(data.antigo, 1, TOL = .01)
fit.atual <- mirt::mirt(data.atual, 1, TOL = .01)

dif <- dif.mirt(fit.antigo = fit.antigo, fit.atual = fit.atual, comuns.atual = comuns.atual, comuns.antigo = comuns.antigo)

}
\author{
Alexandre Jaloto
}
