% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pars.priori.R
\name{pars.priori}
\alias{pars.priori}
\title{Altera priori dos parâmetros}
\usage{
pars.priori(values)
}
\arguments{
\item{values}{tabela de parâmetros gerado pela função \code{mirt} ou \code{multipleGroup} com argumento \code{pars = 'values'}}
}
\value{
Tabela de parâmetros com valores alterados
}
\description{
Altera valores iniciais e distribuição prévia dos parâmetros dos itens na tabela de parâmetros
}
\details{
Distribuição do a é lognormal com média 0 e desvio 0.5,
com valor inicial 1.7. Distribuição do c é beta com alpha 5 e beta 17,
com valor inicial 0.2.
}
\examples{
pars.priori (parametros)

}
\author{
Alexandre Jaloto
}
